+++ src/components/AboutMe/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../styles/AboutMe.module.css'
+import styles from '../../styles/AboutMeSection.module.css'
 
 export default function AboutMe(){
