+++ src/components/Carousel/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../styles/Carousel.module.css'
+import styles from '../../styles/CarouselSection.module.css'
 
 
 export default function Carousel(){
