+++ src/components/Layout/Footer.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../styles/Footer.module.css'
+import styles from '../../styles/FooterMain.module.css'
 
 
 export default function Footer(){
