+++ src/styles/CarouselSection.module.css
@@ -1,0 +1,3 @@
+.carouselSection {
+    width: 100%;
+}
