+++ src/styles/HeaderMain.module.css
@@ -1,4 +1,3 @@
 header {
     background-color: var(--color-grey);
 }
-
