+++ src/styles/NavMenuMain.module.css
@@ -1,10 +1,10 @@
-.titleBurger{
+.titleBurger {
     display: flex;
     justify-content: space-between;
     font-size: 16px;
 }
 
-.burgerMenu{
+.burgerMenu {
     color: var(--color-wheat);
     margin: 10px 20px;
 }
@@ -12,7 +12,7 @@
 .navigation {
     font-weight: bold;
     font-size: 18px;
-    text-transform: uppercase;  
+    text-transform: uppercase;
     color: var(--color-wheat);
     text-align: center;
 }
@@ -25,13 +25,13 @@ li {
     padding: 10px 12px 10px 10px;
     border-style: solid;
     border-width: 1px;
-    }
+}
 
-    a{
-        color: inherit;
-        text-decoration: inherit;
-    }
-    
+a {
+    color: inherit;
+    text-decoration: inherit;
+}
+
 h1 {
     padding: 10px 20px;
     font-weight: bold;
