--- src/styles/carousel.module.css
@@ -1,3 +1,0 @@
-.carouselSection{
-   width: 100%;
-}
